ragsystem/requirements.txt
# Core dependencies
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
html2text>=2024.2.26

# Text processing & chunking
tiktoken>=0.6.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Vector database
pgvector>=0.2.5
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Alternative vector stores (optional)
pinecone-client>=3.0.0
chromadb>=0.4.0

# Web interface
streamlit>=1.31.0
plotly>=5.18.0

# Evaluation & metrics
scikit-learn>=1.4.0
rouge-score>=0.1.2
bert-score>=0.3.13

# Development & testing
pytest>=8.0.0
black>=24.1.0
flake8>=7.0.0

# Async & performance
asyncio>=3.4.3
tqdm>=4.66.0

# Google Cloud (optional)
google-cloud-storage>=2.14.0
google-cloud-aiplatform>=1.42.0
